cub Function to fit a cub model. The structure can have covariates or not to model \(\pi\) and \(\xi\) parameters.

cub

Function to fit a cub model. The structure can have covariates or not to model \(\pi\) and \(\xi\) parameters.

cub(pi.fo, xi.fo, m, data = NULL, subset = NULL,
  optimizer = "nlminb", pi.link = "probit", xi.link = "probit",
  initial.values = NULL, ...)

Arguments

pi.fo
an object of class formula, a symbolic description of the model to fit pi parameter.

xi.fo
an object of class formula, a symbolic description of the model to fit xi parameter without left side.

m
the maximum value of the response variable.

data
an optional data frame.

subset
an optional vector specifying a subset of observations to be used. For example, subset = "sex =='male'" means that we want to use only the male data set.

optimizer
two options are available: nlminb (default), optim or DEoptim.

pi.link
link function to use for model pi parameter, two options are available, logit or probit, by default is probit.

xi.link
link function to use for model xi parameter, two options are available, logit or probit, by default is probit.

initial.values
vector with the initial values to the searching procedure to nlminb and optim optimizers.

...
Further arguments to be passed to DEoptim.control.

Examples

# Test 1 ------------------------------------------------------------------
# Generating a random sample given the values of pi and xi
set.seed(123)
y <- rcub(n=1000, pi=0.15, xi=0.60, m=5)
mod1 <- cub(pi.fo=y ~ 1, xi.fo=~ 1, m=5, optimizer='nlminb')

#> Error in eval(predvars, data, env): objeto 'y' no encontrado

# Summary table
summary(mod1)

#> Error in summary(mod1): objeto 'mod1' no encontrado

# To obtain the fitted parameters in the appropiate scale
# we need to apply the inverse link function to the fitted parameters
pnorm(mod1$par)

#> Error in pnorm(mod1$par): objeto 'mod1' no encontrado

# Using logit link function
mod2 <- cub(pi.fo=y ~ 1, xi.fo=~ 1, m=5, optimizer='nlminb',
            pi.link='logit', xi.link='logit')

#> Error in eval(predvars, data, env): objeto 'y' no encontrado

# Summary table
summary(mod2)

#> Error in summary(mod2): objeto 'mod2' no encontrado

# To obtain the fitted parameters in the appropiate scale we need
# to apply the inverse link function (logit) to the fitted parameters
1 / (1 + exp(-mod2$par))

#> Error in eval(expr, envir, enclos): objeto 'mod2' no encontrado

# Using DEoptim optimizer, it is slower but it gives good results
mod3 <- cub(pi.fo=y ~ 1, xi.fo=~ 1, m=5,
            optimizer='DEoptim', trace=50, itermax=300)

#> Error in eval(predvars, data, env): objeto 'y' no encontrado

# Summary table
summary(mod3)

#> Error in summary(mod3): objeto 'mod3' no encontrado

# we need to apply the inverse link function to obtain the parameters
pnorm(mod3$par)

#> Error in pnorm(mod3$par): objeto 'mod3' no encontrado

# Test 2 ------------------------------------------------------------------
# rcub.covariates is a function to generate a random sample from a cub model 
# using covariates to model pi and xi and link function qnorm
rcub.covariates <- function(n, b0, b1, g0, g1, m = 5) {
  x1 <- runif(n)
  x2 <- runif(n)
  pi <- pnorm(b0 + b1 * x1)
  xi <- pnorm(g0 + g1 * x2)
  y <- rcub(n=n, pi=pi, xi=xi, m=m)
  data.frame(y, x1, x2, xi, pi)
}

# Generating the data
dataset <- rcub.covariates(n=1000, b0=-1, b1=1, g0=-2, g1=1.5)

# Fitting the model with optim optimizer
fit1 <- cub(pi.fo=y ~ x1, xi.fo= ~ x2, m=5, data=dataset,
            optimizer='optim')
summary(fit1)

#> ---------------------------------------------------------------
#> Fixed effects for probit(pi) 
#> ---------------------------------------------------------------
#>             Estimate Std. Error z value  Pr(>|z|)    
#> (Intercept) -0.72680    0.17563 -4.1383 3.499e-05 ***
#> x1           0.67132    0.26333  2.5494   0.01079 *  
#> ---
#> Signif. codes:  0 '***' 0.001 '**' 0.01 '*' 0.05 '.' 0.1 ' ' 1
#> ---------------------------------------------------------------
#> Fixed effects for probit(xi) 
#> ---------------------------------------------------------------
#>             Estimate Std. Error z value  Pr(>|z|)    
#> (Intercept) -1.85375    0.20340 -9.1138 < 2.2e-16 ***
#> x2           1.28115    0.29526  4.3390 1.431e-05 ***
#> ---
#> Signif. codes:  0 '***' 0.001 '**' 0.01 '*' 0.05 '.' 0.1 ' ' 1
#> ---------------------------------------------------------------

# Fitting the model with DEoptim optimizer
fit2 <- cub(pi.fo=y ~ x1, xi.fo=~ x2, m=5, data=dataset,
            optimizer='DEoptim', trace=0, itermax=300)
summary(fit2)

#> ---------------------------------------------------------------
#> Fixed effects for probit(pi) 
#> ---------------------------------------------------------------
#>             Estimate Std. Error z value  Pr(>|z|)    
#> (Intercept) -0.72722    0.17567 -4.1397 3.478e-05 ***
#> x1           0.67181    0.26339  2.5507   0.01075 *  
#> ---
#> Signif. codes:  0 '***' 0.001 '**' 0.01 '*' 0.05 '.' 0.1 ' ' 1
#> ---------------------------------------------------------------
#> Fixed effects for probit(xi) 
#> ---------------------------------------------------------------
#>             Estimate Std. Error z value  Pr(>|z|)    
#> (Intercept) -1.85406    0.20345  -9.113 < 2.2e-16 ***
#> x2           1.28205    0.29534   4.341 1.418e-05 ***
#> ---
#> Signif. codes:  0 '***' 0.001 '**' 0.01 '*' 0.05 '.' 0.1 ' ' 1
#> ---------------------------------------------------------------

# Test 3 ------------------------------------------------------------------
# Simulating a dataset with qualitative variables
n <- 10000
# Betas for pi
b0 <- -1
b1 <- 1
# Betas for xi
g0 <- -2
g1 <- 1.5

beta_japan <- 1 # Coefficient for japan
beta_usa <- -0.7 # Coefficient for usa, france is reference level

# m value
m <- 5

# Simulating the dataset
x1 <- runif(n)
x2 <- runif(n)
country <- sample(c('france', 'japan', 'usa'), size=n, replace=TRUE)
aux <- model.matrix(~country)[, -1] %*% matrix(c(beta_japan,  beta_usa))
pi <- pnorm(b0 + b1 * x1 + aux)
xi <- pnorm(g0 + g1 * x2 + aux)
y <- rcub(n=n, pi=pi, xi=xi, m=m)
# Dataset
dataset <- data.frame(y, x1, x2, country)

# Fitting the model
mod <- cub(pi.fo=y ~ x1 + country,
           xi.fo=~ x2 + country, m=5, data=dataset,
           optimizer='optim')
summary(mod)

#> ---------------------------------------------------------------
#> Fixed effects for probit(pi) 
#> ---------------------------------------------------------------
#>               Estimate Std. Error  z value  Pr(>|z|)    
#> (Intercept)  -1.059794   0.080590 -13.1504 < 2.2e-16 ***
#> x1            1.065660   0.110373   9.6551 < 2.2e-16 ***
#> countryjapan  0.979627   0.077268  12.6784 < 2.2e-16 ***
#> countryusa   -0.664173   0.084222  -7.8860  3.12e-15 ***
#> ---
#> Signif. codes:  0 '***' 0.001 '**' 0.01 '*' 0.05 '.' 0.1 ' ' 1
#> ---------------------------------------------------------------
#> Fixed effects for probit(xi) 
#> ---------------------------------------------------------------
#>               Estimate Std. Error  z value  Pr(>|z|)    
#> (Intercept)  -2.014525   0.067796 -29.7145 < 2.2e-16 ***
#> x2            1.492094   0.063480  23.5049 < 2.2e-16 ***
#> countryjapan  1.014659   0.056240  18.0417 < 2.2e-16 ***
#> countryusa   -0.930035   0.322188  -2.8866  0.003894 ** 
#> ---
#> Signif. codes:  0 '***' 0.001 '**' 0.01 '*' 0.05 '.' 0.1 ' ' 1
#> ---------------------------------------------------------------