cub

function to fit a cub model. The user could use covariates or not to model the parameters pi and xi.

cub(pi.fo, xi.fo, m, data = NULL, subset = NULL,
  optimizer = "nlminb", pi.link = "probit", xi.link = "probit",
  initial.values = NULL, ...)

Arguments

pi.fo
an object of class formula, a symbolic description of the model to fit pi parameter.

xi.fo
an object of class formula, a symbolic description of the model to fit xi parameter without left side.

m
the maximum value of the response variable.

data
an optional data frame.

subset
an optional vector specifying a subset of observations to be used. For example, subset = "sex =='male'" means that we want to use only the male data set.

optimizer
two options are available: nlminb (default), optim or DEoptim.

pi.link
link function to use for model pi parameter, two options are available, logit or probit, by default is probit.

xi.link
link function to use for model xi parameter, two options are available, logit or probit, by default is probit.

initial.values
vector with the initial values to the searching procedure to nlminb and optim optimizers.

...
Further arguments to be passed to DEoptim.control.

Examples

# Example 1 ---------------------------------------------------------------
# Generating a random sample given the values of pi and xi
set.seed(2019)
y <- rcub(n=100, pi=0.15, xi=0.60, m=5)
mod1 <- cub(pi.fo=y ~ 1, xi.fo=~ 1, m=5,
            pi.link='probit', xi.link='probit',
            optimizer='nlminb')

#> Error in eval(predvars, data, env): objeto 'y' no encontrado

summary(mod1)

#> Error in summary(mod1): objeto 'mod1' no encontrado

# estimations for pi and xi
pnorm(mod1$par[1])

#> Error in pnorm(mod1$par[1]): objeto 'mod1' no encontrado

pnorm(mod1$par[2])

#> Error in pnorm(mod1$par[2]): objeto 'mod1' no encontrado

# Example 2 ---------------------------------------------------------------
# Generating a random sample given the values of pi and xi 
# estimation using logit link function
set.seed(2019)
y <- rcub(n=100, pi=0.15, xi=0.60, m=5)
mod2 <- cub(pi.fo=y ~ 1, xi.fo=~ 1, m=5,
            pi.link='logit', xi.link='logit',
            optimizer='nlminb')

#> Error in eval(predvars, data, env): objeto 'y' no encontrado

summary(mod2)

#> Error in summary(mod2): objeto 'mod2' no encontrado

# estimations for pi and xi
boot::inv.logit(mod2$par[1])

#> Error in plogis(x): objeto 'mod2' no encontrado

boot::inv.logit(mod2$par[2])

#> Error in plogis(x): objeto 'mod2' no encontrado

# Example 3 ---------------------------------------------------------------
# Generating a random sample given the values of pi and xi 
# estimation using logit and probit as link function
set.seed(2019)
y <- rcub(n=100, pi=0.15, xi=0.60, m=5)
mod3 <- cub(pi.fo=y ~ 1, xi.fo=~ 1, m=5,
            pi.link='logit', xi.link='probit',
            optimizer='nlminb')

#> Error in eval(predvars, data, env): objeto 'y' no encontrado

summary(mod3)

#> Error in summary(mod3): objeto 'mod3' no encontrado

# estimations for pi and xi
boot::inv.logit(mod3$par[1])

#> Error in plogis(x): objeto 'mod3' no encontrado

pnorm(mod3$par[2])

#> Error in pnorm(mod3$par[2]): objeto 'mod3' no encontrado